requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Admin==1.6.1
Werkzeug==2.3.7
pycryptodome==3.19.0
gunicorn==21.2.0